# docs/swagger.yaml
definitions:
  app.ResponseModel:
    properties:
      data:
        type: object
      message:
        description: Code int         `json:"code"`
        type: string
    type: object
  models.AddMenuForm:
    properties:
      icon_class:
        type: string
      level:
        type: integer
      menu_url:
        type: string
      order_seq:
        type: integer
      parent_menu_id:
        type: integer
      title:
        type: string
    type: object
...
  models.EditMenuForm:
    properties:
      icon_class:
        type: string
      level:
        type: integer
      menu_url:
        type: string
      order_seq:
        type: integer
      parent_menu_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.ForgotForm:
    properties:
      email:
        type: string
    type: object
  models.LoginForm:
    properties:
      p:
        type: string
      u:
        type: string
    type: object
...
  models.RegisterForm:
    properties:
      address:
        type: string
      client_name:
        type: string
      client_type:
        type: string
      contact_person:
        type: string
      created_by:
        type: string
      email_addr:
        type: string
      expiry_date:
        type: string
      joining_date:
        type: string
      post_cd:
        type: string
      start_billing_date:
        type: string
      telephone_no:
        type: string
    type: object
  models.ResetPasswd:
    properties:
      cp:
        type: string
      p:
        type: string
      token_email:
        type: string
    type: object
...
info:
  contact:
    email: nuryantofattih@gmail.com
    name: Nuryanto
    url: https://www.linkedin.com/in/nuryanto-1b2721156/
  description: Backend REST API for golang starter
  license: {}
  title: Starter
  version: "1.0"
paths:
  /api/auth/forgot:
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.ForgotForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      summary: Forgot Password
      tags:
      - Auth
  /api/auth/login:
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.LoginForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      summary: Login
      tags:
      - Auth
  /api/auth/register:
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.RegisterForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      summary: Add Client
      tags:
      - Auth
  /api/auth/reset:
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswd'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      summary: Reset Password
      tags:
      - Auth
  /api/auth/verify:
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswd'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      summary: Verify / Aktivasi User
      tags:
      - Auth
...
  /api/menu:
    get:
      parameters:
      - description: LevelNo
        in: query
        name: level_no
        required: true
        type: integer
      - description: ParentMenuID
        in: query
        name: parent_menu_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: GetList SaMenu
      tags:
      - Menu
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.AddMenuForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Add Menu
      tags:
      - Menu
  /api/menu/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Delete menu
      tags:
      - Menu
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: GetById SaMenu
      tags:
      - Menu
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.EditMenuForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Update Menu
      tags:
      - Menu
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"